requirement.txt
scikit-learn
numpy
pandas
matplotlib
seaborn
streamlit
flask
joblib
openpyxl
mysql-connector-python
